requirements.txt
requests
Pillow
pygame
SpeechRecognition
gtts
picamera
RPi.GPIO
gpiozero
smbus-cffi
httpx[socks]
google-genai